Add Ephemeral storage support for localStorage and sessionStorage
Ephemeral storage persists DOM storage in third-party frames in a
partitioned storage area that is unique to each top-level frame eTLD.
The storage area is persisted until no tabs from the top-level eTLD
exist.

This change also adds a runtime flag for ephemeral storage, which is
only useful if third-party storage is activated. A future change will
selectively enable third-party storage when the flag is active.